+++ changelog.md
@@ -1,5 +1,6 @@
 v2.0.0
 - added !register
+- added !teamcolor
 
 v1.0.2
 - remove challengeme role on quit